@@ -3,22 +3,21 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>ro.victor.training</groupId>
-  <artifactId>jpa</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-  <description>JPA Workshop by Victor Rentea</description>
 
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.4.2</version>
+    <version>3.1.5</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
 
+  <groupId>victor.training</groupId>
+  <artifactId>jpa</artifactId>
+  <version>1.0</version>
+  <packaging>jar</packaging>
+
   <properties>
-    <java.version>11</java.version>
+    <java.version>17</java.version>
   </properties>
 
   <dependencies>
@@ -73,10 +72,11 @@
       <artifactId>querydsl-jpa</artifactId>
     </dependency>
 
-    <dependency>
+   <!-- <dependency>
       <groupId>org.hibernate</groupId>
       <artifactId>hibernate-ehcache</artifactId>
-    </dependency>
+      <version>5.6.9.Final</version>
+    </dependency>-->
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-devtools</artifactId>
@@ -85,12 +85,7 @@
       <dependency>
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
-      <version>2.1.212</version>
-    </dependency>
-    <!-- persist LocalDateTime -->
-    <dependency>
-      <groupId>org.hibernate</groupId>
-      <artifactId>hibernate-java8</artifactId>
+      <version>2.1.214</version>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -113,6 +108,7 @@
       <!--Generate Criteria API metamodel in target/generated-sources/annotations-->
       <groupId>org.hibernate</groupId>
       <artifactId>hibernate-jpamodelgen</artifactId>
+      <version>6.3.1.Final</version>
     </dependency>
   </dependencies>
 
@@ -140,7 +136,11 @@
           </execution>
         </executions>
       </plugin>-->
+
+
+<!--
       <plugin>
+        &lt;!&ndash;QueryDSL metamodel generator&ndash;&gt;
         <groupId>com.mysema.maven</groupId>
         <artifactId>apt-maven-plugin</artifactId>
         <version>1.1.3</version>
@@ -163,6 +163,7 @@
           </dependency>
         </dependencies>
       </plugin>
+-->
     </plugins>
   </build>
 
